+++ flake.nix
@@ -1,12 +1,19 @@
 {
   nixConfig = {
     sandbox = "relaxed";
+    extra-substituters = [
+      "s3://nhost-nix-cache?region=eu-central-1&profile=nhost-nix-cache"
+      "https://cache.nixos.org"
+    ];
+    extra-trusted-public-keys = [
+      "nhost-nix-cache:6bHlSIHLl5ubPXXS0EGgrvEQTyQnc+L05/6vShe/B6g="
+      "cache.nixos.org-1:6NCHdD59X431o0gWypbMrAURkbJ16ZPMQFGspcDShjY="
+    ];
   };
 
   inputs = {
     nixpkgs.url = "github:NixOS/nixpkgs/nixpkgs-unstable";
     flake-utils.url = "github:numtide/flake-utils";
-    nix-filter.url = "github:numtide/nix-filter";
     nix2container.url = "github:nlewo/nix2container";
     nix2container.inputs.nixpkgs.follows = "nixpkgs";
   };
@@ -16,13 +23,12 @@
       self,
       nixpkgs,
       flake-utils,
-      nix-filter,
       nix2container,
     }:
     {
       #nixops
       lib = import ./nixops/lib/lib.nix;
-      overlays.default = import ./nixops/overlays/default.nix { inherit self nix-filter; };
+      overlays.default = import ./nixops/overlays/default.nix;
     }
     // flake-utils.lib.eachDefaultSystem (
       system:
@@ -31,7 +37,7 @@
           inherit system;
           config.allowUnfree = true;
           overlays = [
-            (import ./nixops/overlays/default.nix { inherit self nix-filter; })
+            (import ./nixops/overlays/default.nix)
           ];
         };
 
@@ -42,7 +48,6 @@
           inherit
             self
             pkgs
-            nix-filter
             nixops-lib
             ;
         };
@@ -51,7 +56,6 @@
           inherit
             self
             pkgs
-            nix-filter
             nixops-lib
             ;
         };
@@ -60,7 +64,6 @@
           inherit
             self
             pkgs
-            nix-filter
             nixops-lib
             ;
         };
@@ -69,7 +72,6 @@
           inherit
             self
             pkgs
-            nix-filter
             nixops-lib
             ;
         };
@@ -78,7 +80,6 @@
           inherit
             self
             pkgs
-            nix-filter
             nixops-lib
             ;
         };
@@ -87,7 +88,6 @@
           inherit
             self
             pkgs
-            nix-filter
             nixops-lib
             nix2containerPkgs
             ;
@@ -97,7 +97,6 @@
           inherit
             self
             pkgs
-            nix-filter
             nixops-lib
             nix2containerPkgs
             ;
@@ -108,7 +107,6 @@
             self
             pkgs
             nix2containerPkgs
-            nix-filter
             nixops-lib
             ;
         };
@@ -117,7 +115,6 @@
           inherit
             self
             pkgs
-            nix-filter
             nixops-lib
             ;
         };
@@ -126,7 +123,6 @@
           inherit
             self
             pkgs
-            nix-filter
             nixops-lib
             nix2containerPkgs
             ;
@@ -136,7 +132,6 @@
           inherit
             self
             pkgs
-            nix-filter
             nixops-lib
             ;
         };
@@ -145,7 +140,6 @@
           inherit
             self
             pkgs
-            nix-filter
             nixops-lib
             ;
         };
@@ -154,17 +148,14 @@
           inherit
             self
             pkgs
-            nix-filter
             nixops-lib
             ;
         };
 
         nixopsf = import ./nixops/project.nix {
           inherit
-            self
             pkgs
             nix2containerPkgs
-            nix-filter
             nixops-lib
             ;
         };
@@ -173,7 +164,6 @@
           inherit
             self
             pkgs
-            nix-filter
             nixops-lib
             nix2containerPkgs
             ;
@@ -183,7 +173,6 @@
           inherit
             self
             pkgs
-            nix-filter
             nixops-lib
             ;
         };
@@ -192,7 +181,6 @@
           inherit
             self
             pkgs
-            nix-filter
             nixops-lib
             ;
         };
@@ -201,7 +189,6 @@
           inherit
             self
             pkgs
-            nix-filter
             nixops-lib
             nix2containerPkgs
             ;
@@ -394,6 +381,7 @@
           postgres-pg18-as-dir = postgresf.packages.pg18-as-dir;
           storage = storagef.package;
           storage-docker-image = storagef.dockerImage;
+          storage-vips = storagef.vips;
           clamav-docker-image = storagef.clamav-docker-image;
           tutorials = tutorialsf.package;
         };